WIP: new features in SometimesUnboundedSink for padding
We need the ability to implement _on-demand blocking_ on these
sinks.  (That is, a pair of `set_blocked` and `set_unblocked`
methods.)

While a SometimesUnboundedSink is blocked, it acts as if it's full,
regardless of the state of the underlying sink.  (You can still
use its "unbounded_send" methods.)

Additionally, when a SometimesUnboundedSink is blocked, it has an
upper bound of the number of messages it is willing to flush onto
the underlying sink, even if the underlying sink has room.