proto: Fix a possible misbehavior in SinkBlocker::poll_ready().
`SometimesUnboundedSink` requires us to call poll_ready() on it,
or else it won't progress.  (See comment at the head of that type's
documentation.)

Therefore, since we intend to wrap that type in `SinkBlocker`,
we need `SinkBlocker` to call `inner.poll_ready()` from its own
poll_ready(), even when it is blocking.

I'm hoping that this is a reasonable thing to do, and won't result
in a zillion needless wake-ups.